feat: Update

+++ src/routes.js
@@ -26,6 +26,18 @@ export const routes = [
         handler: (req, res) => {
             const { title, description } = req.body;
 
+            if (!title) {
+                return res.writeHead(400).end(
+                    JSON.stringify({ message: 'title is required' })
+                );
+            }
+
+            if (!description) {
+                return res.writeHead(400).end(
+                    JSON.stringify({ message: 'description is required' })
+                );
+            }
+
             const task = {
                 id: randomUUID(),
                 title,
@@ -41,10 +53,44 @@ export const routes = [
         }
     },
     {
+        method: 'PUT',
+        path: buildRoutePath('/tasks/:id'),
+        handler: (req, res) => {
+            const { id } = req.params;
+            const { title, description } = req.body;
+
+            if (!title && !description) {
+                return res.writeHead(400).end(
+                    JSON.stringify({ message: 'title or description is required' })
+                );
+            }
+
+            const [task] = database.select('tasks', { id });
+
+            if (!task) {
+                return res.writeHead(404).end();
+            }
+
+            database.update('tasks', id, {
+                title: title ?? task.title,
+                description: description ?? task.description,
+                updated_at: new Date()
+            });
+
+            return res.writeHead(204).end();
+        }
+    },
+    {
         method: 'DELETE',
         path: buildRoutePath('/tasks/:id'),
         handler: (req, res) => {
             const { id } = req.params;
+
+            const [task] = database.select('tasks', { id });
+
+            if (!task) {
+                return res.writeHead(404).end();
+            }
 
             database.delete('tasks', id);
